+++ CHANGELOG.md
@@ -29,3 +29,8 @@
 #### Version 1-4-5
 
 * Added optional path selection in write function in Macchiato module
+
+#### Version 1-4-6
+
+* Added support for comments at the end of line in an `*.mpn`
+* Tidied up parsing of command-line arguments added help text